pyproject.toml
[tool.poetry]
name = "infiora-backend"
version = "0.1.0"
description = ""
authors = ["zubair <zubair@xertk.com>"]
license = "MIT"
readme = "README.md"
packages = [{include = "core"}]

[tool.poetry.dependencies]
python = "^3.10"
channels = {extras = ["daphne"], version = "^4.3.1"}
boto3 = "^1.40.32"
django = "^5.2.6"
django-cors-headers = "^4.8.0"
django-filter = "^25.1"
django-model-utils = "^5.0.0"
django-split-settings = "^1.3.2"
django-storages = "^1.14.6"
djangorestframework = "^3.16.1"
djangorestframework-simplejwt = "^5.5.1"
pillow = "^11.3.0"
pydantic = "^2.11.9"
pynacl = "^1.6.0"
whitenoise = "^6.10.0"
pyyaml = "^6.0.2"
psycopg2-binary = "^2.9.10"
python-dotenv = "^1.1.1"

[tool.poetry.group.dev.dependencies]
colorlog = "^6.9.0"
django-debug-toolbar = "^6.0.0"
pre-commit = "^4.3.0"
pytest = "^8.4.2"
pytest-django = "^4.11.1"
pytest-xdist = "^3.8.0"
model-bakery = "^1.20.5"

[tool.isort]
multi_line_output = 5
line_length = 119

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "core.project.settings"
django_find_project = false
python_files = "test_*.py"
testpaths = ["core"]
filterwarnings = "ignore::DeprecationWarning:^(?!node\\.).*:"

[tool.black]
line-length = 119
target-version = ['py310']

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
